+++ datastore/druid/druidDatastoreImpl.go
@@ -518,7 +518,7 @@ type mapLookup struct {
 func (dc *DruidDatastoreClient) UpdateMonitoredObjectMetadata(tenantID string, monitoredObjects []*tenant.MonitoredObject, domains []*tenant.Domain, reset bool) error {
 	version := time.Now().Format(time.RFC3339)
 	lookupEndpoint := dc.coordinatorServer + ":" + dc.coordinatorPort + "/druid/coordinator/v1/lookups/config"
-	lookups := make(map[string]lookup)
+	lookups := make(map[string]lookup, len(domains))
 
 	// Create 1 lookup per domain. Lookups don't support multiple values so the solution is to create
 	// 1 lookup per domain and each lookup has a map where key is monitoredObjectId that belongs in that domain.
@@ -537,15 +537,25 @@ func (dc *DruidDatastoreClient) UpdateMonitoredObjectMetadata(tenantID string, m
 
 	// Fetch existing lookup names and delete any existing lookups on the server that are nolonger valid.
 	// Use the lookup map created in the previous step to identify valid domains.
+	lookupNames := []string{}
 	url := lookupEndpoint + "/__default"
 	result, err := sendRequest("GET", dc.dClient.HttpClient, url, dc.AuthToken, nil)
 	if err != nil {
-		logger.Log.Errorf("Failed to fetch lookups", err.Error())
-		return err
+		if strings.Contains(err.Error(), "No lookups found") {
+			logger.Log.Infof("No lookups found.  Need to initialize lookups before any are created")
+			result, err = sendRequest("POST", dc.dClient.HttpClient, lookupEndpoint, dc.AuthToken, []byte("{}"))
+			if err != nil {
+				logger.Log.Errorf("Failed to initialize druid lookups", err.Error())
+				return err
+			}
+			logger.Log.Infof("Lookups successfully initialized")
+		} else {
+			logger.Log.Errorf("Failed to fetch lookups", err.Error())
+			return err
+		}
+	} else {
+		err = json.Unmarshal(result, &lookupNames)
 	}
-
-	lookupNames := []string{}
-	err = json.Unmarshal(result, &lookupNames)
 
 	// Only delete orphaned domain lookups for this tenant
 	lookupPrefix := buildLookupNamePrefix("dom", tenantID)
@@ -592,6 +602,7 @@ func (dc *DruidDatastoreClient) UpdateMonitoredObjectMetadata(tenantID string, m
 		logger.Log.Errorf("Failed to update lookup", err.Error())
 		return err
 	}
+	refreshLookups(lookups)
 
 	//logger.Log.Debugf("Got result %s", string(result))