+++ datastore/druid/queries.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 
 	pb "github.com/accedian/adh-gather/gathergrpc"
+	"github.com/accedian/adh-gather/logger"
 	"github.com/accedian/godruid"
 )
 
@@ -553,12 +554,20 @@ func buildDomainFilter(tenantID string, domains []string) *godruid.Filter {
 	}
 
 	filters := make([]*godruid.Filter, len(domains))
+	atLeastOneDomainFilter := false
 	for i, domID := range domains {
 		var ef godruid.ExtractionFn
 
+		lookupName, exists := getLookupName("dom", tenantID, domID)
+		if !exists {
+			logger.Log.Warningf("No lookup found for domain ID %s. It will be excluded from the domain filter", domID)
+			continue
+		}
+
+		atLeastOneDomainFilter = true
 		ef = godruid.RegisteredLookupExtractionFn{
 			Type:   "registeredLookup",
-			Lookup: buildLookupName("dom", tenantID, domID),
+			Lookup: lookupName,
 		}
 
 		filters[i] = &godruid.Filter{
@@ -567,6 +576,16 @@ func buildDomainFilter(tenantID string, domains []string) *godruid.Filter {
 			Value:        domID,
 			ExtractionFn: &ef,
 		}
+	}
+
+	if !atLeastOneDomainFilter {
+		// This is a hack to get around no domain lookups being ready yet.  Basically want to
+		// create an 'always false filter".
+		logger.Log.Debugf("No domains found in cached using false filter")
+		return godruid.FilterAnd(
+			godruid.FilterSelector("tenantId", strings.ToLower(tenantID)),
+			godruid.FilterNot(godruid.FilterSelector("tenantId", strings.ToLower(tenantID))),
+		)
 	}
 
 	return godruid.FilterOr(filters...)